fyers_integration/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-103/six-1.15.0-py2.py3-none-any.whl
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.1
